# Searchengine_1/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Спецификация API
  contact:
    email: daniel@gorbe.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.1

servers:
  - url: https://localhost:8080

tags:
  - name: "Индексация"
    description: "Отвечает за индексирование источников информации"
  - name: "Веб-страница"
    description: "Отвечает за индексирование отдельной веб-страницы"
  - name: "Статистика"
    description: "Отвечает за статистику/результат индексации"
  - name: "Поисковый запрос"
    description: "Отвечает за поисковые запросы"

paths:
  /api/startIndexing:
    post:
      tags:
        - "Индексация"
      summary: "Запуск полной индексации"
      description: "Метод запускает полную индексацию всех сайтов или полную
переиндексацию, если они уже проиндексированы.

* Если в настоящий момент индексация или переиндексация уже
запущена, метод возвращает соответствующее сообщение об ошибке."

      responses:
        '200':
          description: "Успешное выполнение запроса"
          content:
            text/plain:
              schema:
                type: boolean
                example: true
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: boolean
                    example: true
        '400':
          description: "Ошибка в выполнении запроса (Bad Request)"
          content:
            text/plain:
              schema:
                type: boolean
                example: false
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Индексация уже запущена"
  /api/stopIndexing:
    post:
      tags:
        - "Индексация"
      summary: "Остановка текущей индексации"
      description: "Метод останавливает текущий процесс индексации (переиндексации).

  * Если в настоящий момент индексация или переиндексация не происходит,
  метод возвращает соответствующее сообщение об ошибке."

      responses:
        '200':
          description: "Успешное выполнение запроса"
          content:
            text/plain:
              schema:
                type: boolean
                example: true
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: boolean
                    example: true
        '400':
          description: "Ошибка в выполнении запроса (Bad Request)"
          content:
            text/plain:
              schema:
                type: boolean
                example: false
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Индексация не запущена"

  /api/indexPage:
    post:
      tags:
        - "Веб-страница"
      summary: "Добавление или обновление отдельной страницы"
      description: "Метод добавляет в индекс или обновляет отдельную страницу, адрес
  которой передан в параметре.

  * Если адрес страницы передан неверно, метод должен вернуть
  соответствующую ошибку."

      parameters:
        - name: url
          in: query
          description: "url — адрес страницы, которую нужно переиндексировать."
          required: false
          schema:
            type: string

      responses:
        '200':
          description: "Успешное выполнение запроса"
          content:
            text/plain:
              schema:
                type: boolean
                example: true
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: boolean
                    example: true

        '400':
          description: "Ошибка в выполнении запроса (Bad Request)"
          content:
            text/plain:
              schema:
                type: boolean
                example: false
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Данная страница находится за пределами сайтов, указанных в конфигурационном файле"

  /api/statistics:
    get:
      tags:
        - "Статистика"
      summary: "Возвращает статистику и результат индексации"
      description: "Метод возвращает статистику и другую служебную информацию о
  состоянии поисковых индексов и самого движка.

  * Если ошибок индексации того или иного сайта нет, задавать ключ error не
  нужно."

      responses:
        '200':
          description: "Успешное выполнение запроса"
          content:
            text/plain:
              schema:
                type: boolean
                example: true
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: boolean
                    example: true
                  statistics:
                    type: object
                    properties:
                      total:
                        type: object
                        properties:
                          sites:
                            type: integer
                            example: 10
                          pages:
                            type: integer
                            example: 1734057
                          lemmas:
                            type: integer
                            example: 512980
                          indexing:
                            type: boolean
                            example: true
                      detailed:
                        type: object
                        properties:
                          url:
                            type: string
                            example: "http://www.site.com"
                          name:
                            type: string
                            example: "Имя сайта"
                          status:
                            type: string
                            example: INDEXED
                          statusTime:
                            type: integer
                            example: 1600160357
                          error:
                            type: string
                            example: "Ошибка индексации: главная страница сайта недоступна"
                          pages:
                            type: integer
                            example: 5678
                          lemmas:
                            type: integer
                            example: 679597
  /api/search:
    get:
      tags:
        - "Поисковый запрос"
      summary: "Получение данных по поисковому запросу"
      description: "Метод осуществляет поиск страниц по переданному поисковому запросу (параметр query). Чтобы выводить результаты порционно, также можно задать параметры offset (сдвиг от начала списка результатов) и limit (количество результатов, которое необходимо вывести). В ответе выводится общее количество результатов (count), не зависящее от значений параметров offset и limit, и массив data с результатами поиска. Каждый результат — это объект, содержащий свойства результата поиска (см. ниже структуру и описание каждого свойства). Если поисковый запрос не задан или ещё нет готового индекса (сайт, по которому ищем, или все сайты сразу не проиндексированы), метод должен вернуть соответствующую ошибку (см. ниже пример). Тексты ошибок должны быть понятными и отражать суть ошибок"

      parameters:

        - name: query
          in: query
          description: 'query — поисковый запрос'
          required: false
          schema:
            type: string

        - name: site
          in: query
          description: 'site — сайт, по которому осуществлять поиск (если не задан, поиск должен происходить по всем проиндексированным сайтам); задаётся в формате адреса, например: http://www.site.com'
          required: false
          schema:
            type: string

        - name: offset
          in: query
          description: 'offset — сдвиг от 0 для постраничного вывода (параметр необязательный; если не установлен, то значение по  умолчанию равно нулю)'
          required: false
          schema:
            type: integer

        - name: limit
          in: query
          description: ' limit — количество результатов, которое необходимо вывести (параметр необязательный; если не установлен, то значение по умолчанию равно 20).'
          required: false
          schema:
            type: integer

      responses:
        '200':
          description: "Успешное выполнение запроса"
          content:
            text/plain:
              schema:
                type: boolean
                example: true

            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: boolean
                    example: true
                  count:
                    type: integer
                    example: 574
                  data:
                    type: object
                    properties:
                      site:
                        type: string
                        example: "http://www.site.com"
                      siteName:
                        type: string
                        example: "Имя сайта"
                      uri:
                        type: string
                        example: "/path/to/page/6784"
                      title:
                        type: string
                        example: "Заголовок страницы, которую выводим"
                      snippet:
                        type: string
                        example: "Фрагмент текста, в которой найдены совпадения, <b> выделенные жирным </b>, в формате HTML"
                      relevance:
                        type: number
                        example: 0.94637

        '400':
          description: "Ошибка в выполнении запроса (Bad Request)"
          content:
            text/plain:
              schema:
                type: boolean
                example: false
            application/json:
              schema:
                type: object
                properties:
                  result:
                    type: boolean
                    example: false
                  error:
                    type: string
                    example: "Задан пустой поисковый запрос"

  /api/uploadDocument:
    post:
      tags: ["Индексация"]
      summary: "Запустить фоновую индексацию PDF/DOCX"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                files:
                  type: array
                  items: {type: string, format: binary}
      responses:
        '202':
          description: "Файлы приняты"
          content:
            application/json:
              schema:
                type: object
                properties:
                  result: {type: boolean}
                  accepted: {type: boolean}
                  jobId: {type: string, format: uuid}

  /api/documents/indexing/status:
    get:
      tags: ["Индексация"]
      summary: "Статус индексации документов текущего пользователя"
      responses:
        '200': {description: "Текущий прогресс и итоговый статус"}

  /api/documents/indexing/stop:
    post:
      tags: ["Индексация"]
      summary: "Остановить индексацию документов текущего пользователя"
      responses:
        '200': {description: "Остановка запрошена"}

  /api/assistant/profile:
    get:
      tags: ["LLM"]
      summary: "Получить профиль RAG текущего пользователя"
      responses:
        '200': {description: "Инструкции и выбранные документы"}
    put:
      tags: ["LLM"]
      summary: "Сохранить профиль RAG текущего пользователя"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                instructions: {type: string, maxLength: 3000}
                documentIds:
                  type: array
                  items: {type: integer}
      responses:
        '200': {description: "Профиль сохранён"}

  /api/assistant/chat:
    post:
      tags: ["LLM"]
      summary: "Ответить по выбранным документам через RAG"
      responses:
        '200': {description: "Ответ и ссылки на источники"}

  /api/assistant/topics:
    get:
      tags: ["LLM"]
      summary: "Определить тематики выбранных документов"
      responses:
        '200': {description: "Темы, уверенность и общий обзор"}